combined entityresourceservice and mapperresourceservice into defaultresourceservice
