+++ src/JsonApiDotNetCore/Extensions/IServiceCollectionExtensions.cs
@@ -93,7 +93,7 @@ namespace JsonApiDotNetCore.Extensions
             if (jsonApiOptions.ContextGraph.UsesDbContext == false)
             {
                 services.AddScoped<DbContext>();
-                services.AddSingleton<DbContextOptions>(new DbContextOptionsBuilder().Options);
+                services.AddSingleton(new DbContextOptionsBuilder().Options);
             }
 
             if (jsonApiOptions.EnableOperations)
@@ -102,29 +102,29 @@ namespace JsonApiDotNetCore.Extensions
             services.AddScoped(typeof(IEntityRepository<>), typeof(DefaultEntityRepository<>));
             services.AddScoped(typeof(IEntityRepository<,>), typeof(DefaultEntityRepository<,>));
 
-            services.AddScoped(typeof(ICreateService<>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(ICreateService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(ICreateService<>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(ICreateService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddScoped(typeof(IGetAllService<>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(IGetAllService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(IGetAllService<>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(IGetAllService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddScoped(typeof(IGetByIdService<>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(IGetByIdService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(IGetByIdService<>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(IGetByIdService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddScoped(typeof(IGetRelationshipService<,>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(IGetRelationshipService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(IGetRelationshipService<,>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(IGetRelationshipService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddScoped(typeof(IUpdateService<>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(IUpdateService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(IUpdateService<>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(IUpdateService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddScoped(typeof(IDeleteService<>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(IDeleteService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(IDeleteService<>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(IDeleteService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddScoped(typeof(IResourceService<>), typeof(EntityResourceService<>));
-            services.AddScoped(typeof(IResourceService<,>), typeof(EntityResourceService<,>));
+            services.AddScoped(typeof(IResourceService<>), typeof(DefaultResourceService<>));
+            services.AddScoped(typeof(IResourceService<,>), typeof(DefaultResourceService<,>));
 
-            services.AddSingleton<JsonApiOptions>(jsonApiOptions);
-            services.AddSingleton<IContextGraph>(jsonApiOptions.ContextGraph);
+            services.AddSingleton(jsonApiOptions);
+            services.AddSingleton(jsonApiOptions.ContextGraph);
             services.AddScoped<IJsonApiContext, JsonApiContext>();
             services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
             services.AddScoped<IScopedServiceProvider, RequestScopedServiceProvider>();
